Add files via upload
A brief understanding of Hr Industry Human resources (HR) is the division of an organization that is responsible finding, screening, recruiting and training job applicants. And administrating the employee benifit programmes. The management is the strategic apporach to the efficient management of people in any organization in that can help their business gain a competitive advantage. It is designrd to improve employee performance in service of employee staregic objectives.